+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_server.py
@@ -30,6 +30,7 @@ import grpc
 
 from opentelemetry import propagators, trace
 from opentelemetry.context import attach, detach
+from opentelemetry.trace.propagation.textmap import Getter
 
 from . import grpcext
 from ._utilities import RpcInfo
@@ -110,18 +111,6 @@ def _check_error_code(span, servicer_context, rpc_info):
         rpc_info.error = servicer_context.code
 
 
-class Getter:
-    @staticmethod
-    def get(metadata, key) -> List[str]:
-        md_dict = {md.key: md.value for md in metadata}
-        return [md_dict[key]] if key in md_dict else []
-
-    @staticmethod
-    def keys(metadata) -> List[str]:
-        md_dict = {md.key: md.value for md in metadata}
-        return md_dict.keys()
-
-
 class OpenTelemetryServerInterceptor(
     grpcext.UnaryServerInterceptor, grpcext.StreamServerInterceptor
 ):
@@ -133,10 +122,14 @@ class OpenTelemetryServerInterceptor(
     def _set_remote_context(self, servicer_context):
         metadata = servicer_context.invocation_metadata()
         if metadata:
+            md_dict = {md.key: md.value for md in metadata}
 
-            get_from_grpc_metadata = Getter()
+            def get_from_grpc_metadata(metadata, key) -> List[str]:
+                return [md_dict[key]] if key in md_dict else []
+
+            getter = Getter(get_from_grpc_metadata)
             # Update the context with the traceparent from the RPC metadata.
-            ctx = propagators.extract(get_from_grpc_metadata, metadata)
+            ctx = propagators.extract(getter, metadata)
             token = attach(ctx)
             try:
                 yield