+++ instrumentation/opentelemetry-instrumentation-opentracing-shim/src/opentelemetry/instrumentation/opentracing_shim/__init__.py
@@ -112,6 +112,7 @@ from opentelemetry.trace import (
     get_current_span,
     set_span_in_context,
 )
+from opentelemetry.trace.propagation.textmap import Getter
 from opentelemetry.util.types import Attributes
 
 ValueT = TypeVar("ValueT", int, float, bool, str)
@@ -503,47 +504,6 @@ class ScopeManagerShim(ScopeManager):
         return self._tracer
 
 
-class Getter:
-    """This class provides an interface that enables extracting propagated
-    fields from a carrier
-
-    """
-
-    @staticmethod
-    def get(carrier, key):
-        """Function that can retrieve zero
-        or more values from the carrier. In the case that
-        the value does not exist, returns an empty list.
-
-        Args:
-            carrier: and object which contains values that are
-                used to construct a Context. This object
-                must be paired with an appropriate get_from_carrier
-                which understands how to extract a value from it.
-            key: key of a field in carrier.
-        Returns:
-            first value of the propagation key or an empty list if the key doesn't exist.
-        """
-
-        value = carrier.get(key)
-        return [value] if value is not None else []
-
-    @staticmethod
-    def keys(carrier):
-        """Function that can retrieve all the keys in a carrier object.
-
-        Args:
-            carrier: and object which contains values that are
-                used to construct a Context. This object
-                must be paired with an appropriate get_from_carrier
-                which understands how to extract a value from it.
-        Returns:
-            list of keys from the carrier.
-        """
-
-        return carrier.keys()
-
-
 class TracerShim(Tracer):
     """Wraps a :class:`opentelemetry.trace.Tracer` object.
 
@@ -747,9 +707,14 @@ class TracerShim(Tracer):
         if format not in self._supported_formats:
             raise UnsupportedFormatException
 
-        get_as_list = Getter()
+        def get_as_list(dict_object, key):
+            value = dict_object.get(key)
+            return [value] if value is not None else []
+
+        getter = Getter(get_as_list)
+
         propagator = propagators.get_global_textmap()
-        ctx = propagator.extract(get_as_list, carrier)
+        ctx = propagator.extract(getter, carrier)
         span = get_current_span(ctx)
         if span is not None:
             otel_context = span.get_context()