+++ instrumentation/opentelemetry-instrumentation-tornado/src/opentelemetry/instrumentation/tornado/__init__.py
@@ -53,6 +53,7 @@ from opentelemetry.instrumentation.utils import (
     http_status_to_canonical_code,
     unwrap,
 )
+from opentelemetry.trace.propagation.textmap import Getter
 from opentelemetry.trace.status import Status
 from opentelemetry.util import ExcludeList, time_ns
 
@@ -174,15 +175,11 @@ def _log_exception(tracer, func, handler, args, kwargs):
     return func(*args, **kwargs)
 
 
-class Getter:
-    @staticmethod
-    def get(headers: dict, header_name: str) -> typing.List[str]:
-        header = headers.get(header_name)
-        return [header] if header else []
-
-    @staticmethod
-    def keys(headers) -> typing.List[str]:
-        return headers.keys()
+def _get_header_from_request_headers(
+    headers: dict, header_name: str
+) -> typing.List[str]:
+    header = headers.get(header_name)
+    return [header] if header else []
 
 
 def _get_attributes_from_request(request):
@@ -210,11 +207,9 @@ def _get_operation_name(handler, request):
 
 
 def _start_span(tracer, handler, start_time) -> _TraceContext:
-    _get_header_from_request_headers = Getter()
+    getter = Getter(_get_header_from_request_headers)
     token = context.attach(
-        propagators.extract(
-            _get_header_from_request_headers, handler.request.headers,
-        )
+        propagators.extract(getter, handler.request.headers,)
     )
     span = tracer.start_span(
         _get_operation_name(handler, handler.request),