+++ opentelemetry-sdk/tests/trace/propagation/test_b3_format.py
@@ -19,27 +19,17 @@ import opentelemetry.sdk.trace as trace
 import opentelemetry.sdk.trace.propagation.b3_format as b3_format
 import opentelemetry.trace as trace_api
 from opentelemetry.context import get_current
+from opentelemetry.trace.propagation.textmap import Getter
 
 FORMAT = b3_format.B3Format()
 
 
-class Getter:
-    @staticmethod
-    def get(dict_object, key):
-        value = dict_object.get(key)
-        return [value] if value is not None else []
-
-    @staticmethod
-    def keys(dict_object):
-        return dict_object.keys()
-
-
-get_as_list = Getter()
+getter = Getter()
 
 
 def get_child_parent_new_carrier(old_carrier):
 
-    ctx = FORMAT.extract(get_as_list, old_carrier)
+    ctx = FORMAT.extract(getter, old_carrier)
     parent_context = trace_api.get_current_span(ctx).get_context()
 
     parent = trace.Span("parent", parent_context)
@@ -240,7 +230,7 @@ class TestB3Format(unittest.TestCase):
         invalid SpanContext.
         """
         carrier = {FORMAT.SINGLE_HEADER_KEY: "0-1-2-3-4-5-6-7"}
-        ctx = FORMAT.extract(get_as_list, carrier)
+        ctx = FORMAT.extract(getter, carrier)
         span_context = trace_api.get_current_span(ctx).get_context()
         self.assertEqual(span_context.trace_id, trace_api.INVALID_TRACE_ID)
         self.assertEqual(span_context.span_id, trace_api.INVALID_SPAN_ID)
@@ -252,7 +242,7 @@ class TestB3Format(unittest.TestCase):
             FORMAT.FLAGS_KEY: "1",
         }
 
-        ctx = FORMAT.extract(get_as_list, carrier)
+        ctx = FORMAT.extract(getter, carrier)
         span_context = trace_api.get_current_span(ctx).get_context()
         self.assertEqual(span_context.trace_id, trace_api.INVALID_TRACE_ID)
 
@@ -276,7 +266,7 @@ class TestB3Format(unittest.TestCase):
             FORMAT.FLAGS_KEY: "1",
         }
 
-        ctx = FORMAT.extract(get_as_list, carrier)
+        ctx = FORMAT.extract(getter, carrier)
         span_context = trace_api.get_current_span(ctx).get_context()
 
         self.assertEqual(span_context.trace_id, 1)
@@ -302,7 +292,7 @@ class TestB3Format(unittest.TestCase):
             FORMAT.FLAGS_KEY: "1",
         }
 
-        ctx = FORMAT.extract(get_as_list, carrier)
+        ctx = FORMAT.extract(getter, carrier)
         span_context = trace_api.get_current_span(ctx).get_context()
 
         self.assertEqual(span_context.trace_id, 1)
@@ -315,7 +305,7 @@ class TestB3Format(unittest.TestCase):
             FORMAT.FLAGS_KEY: "1",
         }
 
-        ctx = FORMAT.extract(get_as_list, carrier)
+        ctx = FORMAT.extract(getter, carrier)
         span_context = trace_api.get_current_span(ctx).get_context()
         self.assertEqual(span_context.span_id, trace_api.INVALID_SPAN_ID)
 
@@ -333,5 +323,5 @@ class TestB3Format(unittest.TestCase):
         def setter(carrier, key, value):
             carrier[key] = value
 
-        ctx = FORMAT.extract(get_as_list, {})
+        ctx = FORMAT.extract(getter, {})
         FORMAT.inject(setter, {}, ctx)